Direct3D 9 conversion
Only 32 bit graphics
  No RGBA masks, no rgb_packed
  SCREEN data is unsigned*, pitch is in DWORDs, no pixel_length
  Excluded many non-32 bit pixel code (drawing functions)
  No palettes

PSD graphics are unpacked into 32-bit RGB (only RGBX and RGBA formats) (RGB_DATA, ref couunted)
Cockpit graphics are not unpacked (ex-RLE)

Old pictured 640x480x16 cockpits are not supported
When custom cockpit is present old code is commented

Not use special memory allocators

TEXTURE_MAP_TYPES replaced with TEXTURE_ROLE

Fixed some memory problems

Virtual FOV

Problems: Havoc cockpit needles/RWR, MFD export requires 3D device.

+++ aphavoc/source/entity/mobile/aircraft/fixwing/fw_updt.c
@@ -199,7 +199,7 @@ static void update_server (entity *en)
 		}
 		
 		if (damage_type > AIRCRAFT_DAMAGE_NONE)
-			if (frand1() < pow(damage_type, 4) * get_delta_time())
+			if (frand1() < pow((float)damage_type, 4) * get_delta_time())
 			{
 				int damage_level = max(0, get_local_entity_int_value (en, INT_TYPE_DAMAGE_LEVEL) - 1);
